+++ src/actions/skill-action.ts
@@ -2,8 +2,18 @@
 import { Skill, SkillSchema } from "@/lib/models/skill";
 import { addSkill, deleteSkill, getSkills, updateSkill, getActiveSkills } from "@/lib/services/skills-service";
 import { revalidatePath } from "next/cache";
-
+import { checkAuth } from "@/lib/auth";
+import { cookies } from "next/headers";
 export async function addSkillAction(skill: Skill) {
+    const cookieStore = await cookies();
+    const token = cookieStore.get('auth_code')?.value;
+    if (!token) {
+        return { success: false, message: "Unauthorized", status: 401 };
+    }
+    const auth = await checkAuth(token);
+    if (!auth) {
+        return { success: false, message: "Unauthorized", status: 401 };
+    }
     const validatedSkill = SkillSchema.safeParse(skill);
     if (!validatedSkill.success) {
         throw new Error(validatedSkill.error.message);
@@ -25,6 +35,15 @@ export async function getActiveSkillsAction() {
     }
 }
 export async function updateSkillAction(skill: Skill) {
+    const cookieStore = await cookies();
+    const token = cookieStore.get('auth_code')?.value;
+    if (!token) {
+        return { success: false, message: "Unauthorized", status: 401 };
+    }
+    const auth = await checkAuth(token);
+    if (!auth) {
+        return { success: false, message: "Unauthorized", status: 401 };
+    }
     const validatedSkill = SkillSchema.safeParse(skill);
     if (!validatedSkill.success) {
         throw new Error(validatedSkill.error.message);
@@ -38,6 +57,15 @@ export async function updateSkillAction(skill: Skill) {
     }
 }
 export async function deleteSkillAction(id: number) {
+    const cookieStore = await cookies();
+    const token = cookieStore.get('auth_code')?.value;
+    if (!token) {
+        return { success: false, message: "Unauthorized", status: 401 };
+    }
+    const auth = await checkAuth(token);
+    if (!auth) {
+        return { success: false, message: "Unauthorized", status: 401 };
+    }
     try {
         revalidatePath("/");
         return await deleteSkill(id);
@@ -47,6 +75,15 @@ export async function deleteSkillAction(id: number) {
     }
 }
 export async function getSkillsAction() {
+    const cookieStore = await cookies();
+    const token = cookieStore.get('auth_code')?.value;
+    if (!token) {
+        return { success: false, message: "Unauthorized", status: 401 };
+    }
+    const auth = await checkAuth(token);
+    if (!auth) {
+        return { success: false, message: "Unauthorized", status: 401 };
+    }
     try {
         return await getSkills();
     } catch (error) {
